adopting unit tests for new error types
