+++ src/xcode/ENA/ENA/Source/HTTPClient/Resources/ReceiveResources/_ReceiveResource.swift
@@ -2,6 +2,20 @@
 // 🦠 Corona-Warn-App
 //
 
+import HealthCertificateToolkit
+
+/**
+ Errors for decoding cases when we get a specialized error from the decoding
+ */
+enum ModelDecodingError: Error {
+	case STRING_DECODING
+	case PROTOBUF_DECODING(Error)
+	case JSON_DECODING(Error)
+	case CBOR_DECODING
+	case CBOR_DECODING_VALIDATION_RULES(RuleValidationError)
+	case CBOR_DECODING_ONBOARDED_COUNTRIES(RuleValidationError)
+}
+
 /**
 A ReceiveResource knows how to decode the data of the http response body. At the end, we receive a concrete object (for example an JSON Object or a protbuf).
 The resource only knows of which type ReceiveModel is and implements the concrete encode function to get at the end a concrete object of the http response body's data.
