+++ src/xcode/ENA/ENA/Source/HTTPClient/Resources/__tests__/DCCRulesResourceTests.swift
@@ -123,10 +123,11 @@ final class DCCRulesResourceTests: CWATestCase {
 				XCTFail("Load should fail but failed succeeded 😅")
 			case let .failure(error):
 				// THEN
-				// Successful test if we can unpack the error to an .ONBOARDED_COUNTRIES_DECODING_ERROR containing a .CBOR_DECODING_FAILED error.
+				// Successful test if we can unpack the error to an RuleValidationError.CBOR_DECODING_FAILED containing a some error we are not interested in.
 				guard case let .receivedResourceError(customError) = error,
 					  case let .RULE_DECODING_ERROR(.acceptance, decodingError) = customError,
-					  case .CBOR_DECODING_FAILED = decodingError else {
+					  case let .CBOR_DECODING_VALIDATION_RULES(ruleValidationError) = decodingError,
+					  case RuleValidationError.CBOR_DECODING_FAILED = ruleValidationError else {
 						  XCTFail("Received wrong error type")
 						  return
 				}
