+++ src/xcode/ENA/ENA/Source/HTTPClient/Resources/__tests__/RegistrationTokenResourceTests.swift
@@ -89,7 +89,9 @@ final class RegistrationTokenResourceTests: CWATestCase {
 			case .failure(let error):
 				// Successful test if we can extract the resourceError to an decoding error, regardless what JSON decoding error exactly it is.
 				guard case let .resourceError(resourceError) = error,
-					  case _ = ResourceError.decoding(resourceError) else {
+					  case let .decoding(decodingError) = resourceError,
+					  case let .JSON_DECODING(someError) = decodingError,
+					  case _ = someError else {
 						 
 					XCTFail("unexpected error case")
 					return
@@ -124,7 +126,9 @@ final class RegistrationTokenResourceTests: CWATestCase {
 			case .failure(let error):
 				// Successful test if we can extract the resourceError to an decoding error, regardless what JSON decoding error exactly it is.
 				guard case let .resourceError(resourceError) = error,
-					  case _ = ResourceError.decoding(resourceError) else {
+					  case let .decoding(decodingError) = resourceError,
+					  case let .JSON_DECODING(someError) = decodingError,
+					  case _ = someError else {
 						 
 					XCTFail("unexpected error case")
 					return
