+++ src/xcode/ENA/ENA/Source/HTTPClient/Resources/__tests__/TeleTanResourceTests.swift
@@ -199,7 +199,8 @@ final class TeleTanResourceTests: CWATestCase {
 			case .failure(let error):
 				// Successful test if we can extract the resourceError to an decoding error, regardless what JSON decoding error exactly it is.
 				guard case let .resourceError(resourceError) = error,
-					  case _ = ResourceError.decoding(resourceError) else {
+					  case let .decoding(decodingError) = resourceError,
+					  case .JSON_DECODING = decodingError else {
 						 
 					XCTFail("unexpected error case")
 					return
@@ -236,7 +237,9 @@ final class TeleTanResourceTests: CWATestCase {
 			case .failure(let error):
 				// Successful test if we can extract the resourceError to an decoding error, regardless what JSON decoding error exactly it is.
 				guard case let .resourceError(resourceError) = error,
-					  case _ = ResourceError.decoding(resourceError) else {
+					  case let .decoding(decodingError) = resourceError,
+					  case let .JSON_DECODING(someError) = decodingError,
+					  case _ = someError else {
 						 
 					XCTFail("unexpected error case")
 					return
