+++ src/xcode/ENA/ENA/Source/HTTPClient/Service/Models/Receive/ValidationRulesModel.swift
@@ -7,16 +7,6 @@ import Foundation
 import HealthCertificateToolkit
 import class CertLogic.Rule
 
-enum ModelDecodingError: Error {
-	case STRING_DECODING
-	case PROTOBUF_DECODING(Error)
-	case JSON_DECODING(Error)
-	case CBOR_DECODING
-
-	case CBOR_DECODING_VALIDATION_RULES(RuleValidationError)
-	case CBOR_DECODING_ONBOARDED_COUNTRIES(RuleValidationError)
-}
-
 struct ValidationRulesModel: CBORDecoding {
 
 	static func decode(_ data: Data) -> Result<ValidationRulesModel, ModelDecodingError> {
